README.md
This program implements priority scheduling algorithm with context switching time. 
The user is Prompted to enter the number of processes and then enter their priority, burst time and arrival time also. 
Whenever operating system preempts a process and shifts CPU's control to some another process of higher priority assuming that it takes 2 seconds for context switching(dispatcher latency).
A scenario is formed, where we can give the processes are assigned with priority where the lower integer number is higher priority and then context switch .. as the process waits the priority of the process increase at rate of one per 2 time units of wait. The waiting time and turnaround time for each process is calculated and printed on the screen.